HF_JQuery Ch.11 Finished BOOK FINISHED
Need to fix problem with ch. 9 code and to upload and try out ch. 10,
ch. 11 code and then you are truely done. The in order to make it stick
go through every chapter and start a mini-project in order to have some
more practice and a brief overview of what was learned